[CSL-778] Move base part of FTS to lrc package
