+++ lrc/cardano-sl-lrc.cabal
@@ -12,24 +12,31 @@ build-type:          Simple
 cabal-version:       >=1.10
 
 library
-  exposed-modules:
+  exposed-modules:     Pos.Lrc.FollowTheSatoshiM
 
   build-depends:       base
                      , bytestring
-                     -- , cardano-sl-core
-                     -- , concurrent-extra
-                     -- , data-default
-                     -- , formatting
-                     -- , lens
-                     -- , log-warper
-                     -- , mtl
-                     -- , node-sketch
-                     -- , resourcet
-                     -- , rocksdb
-                     -- , serokell-util
-                     -- , transformers-base
-                     -- , text-format
-                     -- , universum
+                     , unordered-containers
+                     , cardano-sl-core
+                     , universum
+
+  default-extensions:  DeriveDataTypeable
+                       DeriveGeneric
+                       GeneralizedNewtypeDeriving
+                       StandaloneDeriving
+                       FlexibleContexts
+                       FlexibleInstances
+                       MultiParamTypeClasses
+                       FunctionalDependencies
+                       DefaultSignatures
+                       NoImplicitPrelude
+                       OverloadedStrings
+                       RecordWildCards
+                       TypeApplications
+                       TupleSections
+                       ViewPatterns
+                       LambdaCase
+                       MultiWayIf
 
   default-language:    Haskell2010
